Replace Intervener with ActiveIntervener and MuteIntervener
This commit replaces the Intervener in timeinterval package with
an ActiveIntervener and a MuteIntervener for active time intervals
and mute time intervals. It avoids having a double-negative in
TimeActiveStage where Mute returns true if its not muted rather
than if it is muted, and will also enable us to mark alerts as
muted or not within ActiveIntervener and MuteIntervener.

Signed-off-by: George Robinson <george.robinson@grafana.com>