Log an error if the parsers produce different parsings
This commit updates the fallback parser to compare the parsings
of both the UTF-8 matchers parser and the classic parser, and
log an error if the parsers produce different parsings.

This can happen at present if ecsape sequences such as \n and \t
are used outside of double quotes, as this is not supported in
the UTF-8 matchers parser.

Signed-off-by: George Robinson <george.robinson@grafana.com>