Allow expiration of invalid silences
A silence can become invalid if created with the UTF-8 matchers
parser, but deleted when Alertmanager is running with the classic
matchers parser feature flag.

Without this change, deletion of such silences fail as the silence
is considered invalid, forcing the silence to remain active until
either it's expiration time elapses, or the Alertmanager is restarted
without the feature flag and the operation repeated.

Signed-off-by: George Robinson <george.robinson@grafana.com>