Mark the fingerprints of all inhibiting rules
Before this change, Alertmanager would mark just the fingerprint of
the first inhibition rule to inhibit an alert, rather than the
fingerprints of all inhibition rules that inhibit the alert.
This is different from silences which mark the IDs of all silences
that silence an alert, not just the first silence.

This commit changes how inhibition rules work to match the
behavior of silences, such that the fingerprints of all inhibition
rules that inhibit the alert are marked.

This change is expected to increase CPU usage as the inhibiter has to
do more work, matching the label set against all inhibition rules
instead of stopping at the first match. However, this is no different
from how silences mute alerts.

Signed-off-by: George Robinson <george.robinson@grafana.com>