Make route.Key() unique and remove route.ID()
This commit replaces the code in route.Key() with that of route.ID(),
and removes route.ID(). The motivation behind this change is to fix
a number of bugs caused by conflicting group keys such as
"Different aggregation groups can share the same nflog" (#3808)
and also prevent an issue where groups are incorrectly marked as muted
when they are not.

Signed-off-by: George Robinson <george.robinson@grafana.com>