Replace DeleteIf with DeleteIfNotModified
This commit replaces DeleteIf with DeleteIfNotModified as this
allows dispatch.go to delete all resolved alerts in a single
atomic operation instead of having to acquire and release
the store mutex one alert at a time.

It also extends the test coverage in dispatch_test.go to make
sure that just the resolved alerts are deleted from the
aggregation group on flush.

Signed-off-by: George Robinson <george.robinson@grafana.com>