Remove GC and callback from store.go
This commit removes the GC and callback function from store.go
to address a number of data races that have occurred in the past
(#2040 and #3648). The store is no longer responsible for removing
resolved alerts after some elapsed period of time, and is instead
deferred to the consumer of the store (as done in #2040 and #3648).

Signed-off-by: George Robinson <george.robinson@grafana.com>