Support Noop limits for silences
This commit changes how limits work for silences. Instead of checking
if a limit is non-nil, it instead adds support for a set of no-op
limits called NoopLimits.

It changes the limits in Options and Silences from Limits to *Limits
because Go cannot do comparisons on structs containing function
values, so we must use a nil check to see if limits are set in the
options or not.

Signed-off-by: George Robinson <george.robinson@grafana.com>