Fix subtle bug in JSON/YAML encoding of inhibition rules
This commit fixes a subtle bug introduced in #4177 where external
code that imports config.InhibitRule for the purpose of JSON/YAML
encoding can encode an inhibition rule with missing equals labels.
This bug will happen if someone updates their Go modules, writes
labels to the Equal field, and does not also write the same labels
to the EqualStr field.

To fix this, I propose removing EqualStr and changing Equal from
model.LabelNames to []string. This will cause an intentional
breaking change, and require the maintainer of downstream code to
use []string instead of model.LabelNames. I consider this a better
option to the subtle bug explained earlier.

This still keeps the fix for #4177, but does so with an explicit
breaking change to downstream users.

Signed-off-by: George Robinson <george.robinson@grafana.com>