fix(ui usability): Quote matcher values when viewing silences.
Problem: When viewing a silence in the UI (in the list or individually)
its matchers are rendered without quotes around the value, which makes
it impossible to differentiate between matchers like `a="~b"` and
`a=~"b"`. That makes silences difficult to debug using the web UI.

Fix: Quote the value of the matcher, just as it is quoted when
editing/adding matchers in a silence.

Signed-off-by: Rudolf Thomas <rudolf.thomas@firma.seznam.cz>