Add metrics for the case where the matches fails to compile/add to the cache
Before the silences would be in the status, but querying would result in
query errors. Now they are still in the status (since we have to
calculate after calling Merge), but we also know that some broken ones
exist.

At load snapshot time we can also avoid importing those, since they
would not be usable anyway, and we have different labels to distinguish
the cases/behaviors.

Signed-off-by: Guido Trotter <guido@hudson-trading.com>