Fix more flaky tests in TestClusterJoinAndReconnect (#4701)
As explained in more detail in commit
`88b4e13d09975b61290493f8893bb4fb7796a93f`, the tests in
`cluster_test.go` will occasionally fail. The underlying reason for the
failure is that the test only waits for `p2` to be ready, but this does
not reflect whether `p1` has updated its memberlist.

However, that commit only fixed the test `TestJoinLeave`. I have now ran
the other tests in `cluster_test.go` 2400 times, and identified the
following problematic tests: `TestSetPeerNames` and `TestTLSConnection`.

Since the underlying faulty code is mostly identical, so is the
corresponding fix.

Signed-off-by: Solomon Jacobs <solomonjacobs@protonmail.com>