Logging improvements (#4113)
* ref(webhook): return error if unable to parse group key

Addresses feedback from another PR
https://github.com/prometheus/alertmanager/pull/4089#discussion_r1824480384

Signed-off-by: TJ Hoplock <t.hoplock@gmail.com>

* fix(silence): log errors at ERROR level instead of INFO

Signed-off-by: TJ Hoplock <t.hoplock@gmail.com>

* chore(notify): make logging more consistent, converge on `group_key`

This changes the majority of our `Notify()` implementations to set up a
new logger with the group key attached under the key name `group_key`,
and then to use that logger in all subsequent calls to the logger,
including passing it through to further functions that accept loggers
via params.

A few of the notify implementations are more complicated; they either
extract the key later in their `Notify()` implementation or within sub
methods, or even conditionally like with sns. I left those mostly as is
for now, as they seem to be more snow-flake-y.

Signed-off-by: TJ Hoplock <t.hoplock@gmail.com>

---------

Signed-off-by: TJ Hoplock <t.hoplock@gmail.com>